Done all tasks

+++ 09 Приведение примитивных типов расширение и сужение/task1008/src/task1008/Solution.java
@@ -1,3 +1,5 @@
+//Complete
+
 package task1008;
 
 /* 
@@ -21,7 +23,7 @@ public class Solution {
     public static void main(String[] args) {
         int a = 50;
         int b = 17;
-        double d = (float) a / b;
+        double d = (double) a / b;
         System.out.println(d);
     }
 }
